Post 조회시 이미지 객체 포함해서 리턴 글, 댓글 삭제할 때 사용자 유효성 검사 추가

+++ backend/src/main/java/com/web/blog/controller/post/CommentController.java
@@ -111,10 +111,15 @@ public class CommentController {
     @DeleteMapping("/{cid}")
     @ApiOperation(value = "글 삭제")
 
-    public Object delete(@Valid @PathVariable int cid) {
+    public Object delete(@Valid @PathVariable int cid, @RequestBody String jsonObj) throws ParseException {
         Comment findComment = commentRepository.getCommentByCid(cid);
         BasicResponse result = new BasicResponse();
-        if(findComment != null){
+        JSONParser jsonParse = new JSONParser();
+        JSONObject obj = (JSONObject) jsonParse.parse(jsonObj);
+
+        String token = (String) obj.get("token");
+        String name = commentRepository.getCommentByCid(cid).getName();
+        if((findComment != null) && (tokenUtils.getUserNameFromToken(token).equals(name))){
             commentRepository.delete(findComment);
             result.data = SUCCESS;
             result.status = true;
